requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Flask-CORS==4.0.0
Werkzeug==2.3.7
requests==2.31.0
zhipuai==2.0.1
openai==1.58.1
PyJWT==2.8.0
numpy
PyPDF2==3.0.1
python-pptx==0.6.21
python-docx==0.8.11